+++ tests/notebooks/mirror/ipynb_to_script_vscode_folding_markers/nteract_with_parameter.py
@@ -10,21 +10,21 @@
 #     name: python3
 # ---
 
-# region outputHidden=false inputHidden=false tags=["parameters"]
+# region inputHidden=false outputHidden=false tags=["parameters"]
 param = 4
 # endregion
 
-# region outputHidden=false inputHidden=false
+# region inputHidden=false outputHidden=false
 import pandas as pd
 # endregion
 
-# region outputHidden=false inputHidden=false
+# region inputHidden=false outputHidden=false
 df = pd.DataFrame({'A': [1, 2], 'B': [3 + param, 4]},
                   index=pd.Index(['x0', 'x1'], name='x'))
 df
 # endregion
 
-# region outputHidden=false inputHidden=false
+# region inputHidden=false outputHidden=false
 # %matplotlib inline
 df.plot(kind='bar')
 # endregion